refactor: utils json fectch 모듈화
