+++ utils.js
@@ -10,4 +10,32 @@ function showToday() {
     });
 }
 
+async function fetchPressData() {
+    try {
+        const data = await fetch("./data/press_data.json")
+            .then((res) => res.json())
+            .then((data) => data.sort(() => Math.random() - 0.5));
+        // then return data
+        return data;
+    } catch (error) {
+        console.log(error);
+        return null;
+    }
+}
+
+async function fetchHotTopicData() {
+    try {
+        const data = await fetch("./data/hot_topic_data.json")
+            .then((res) => res.json())
+            .then((data) => data.sort(() => Math.random() - 0.5));
+        // then return data
+        return data;
+    } catch (error) {
+        console.log(error);
+        return null;
+    }
+}
+
 document.addEventListener("DOMContentLoaded", showToday);
+
+export { fetchPressData, fetchHotTopicData };
